refactor: split server util files
Gives better tree shaking because now game server wont import stuff only used by API and vice-versa

will also make splitting API and Game server into their own packages easier in the future
