+++ server/src/api/auth/middleware.ts
@@ -3,7 +3,8 @@ import type { Context, Next } from "hono";
 import { deleteCookie, getCookie, setCookie } from "hono/cookie";
 import type { z } from "zod";
 import { Config } from "../../config.ts";
-import { getHonoIp, HTTPRateLimit } from "../../utils/serverHelpers.ts";
+import { HTTPRateLimit } from "../../utils/rateLimit.ts";
+import { getHonoIp } from "../apiHelpers.ts";
 import { server } from "../apiServer.ts";
 import { cookieDomain, deleteSessionTokenCookie } from "../routes/user/auth/authUtils.ts";
 import { validateSessionToken } from "./index.ts";
