+++ server/src/api/routes/user/auth/authUtils.ts
@@ -6,7 +6,7 @@ import slugify from "slugify";
 import { UnlockDefs } from "../../../../../../shared/defs/gameObjects/unlockDefs.ts";
 import { util } from "../../../../../../shared/utils/util.ts";
 import { Config } from "../../../../config.ts";
-import { checkForBadWords } from "../../../../utils/serverHelpers.ts";
+import { checkForBadWords } from "../../../../utils/badWords.ts";
 import { createSession, invalidateSession } from "../../../auth/index.ts";
 import { db } from "../../../db/index.ts";
 import { itemsTable, usersTable, type UsersTableInsert } from "../../../db/schema.ts";
