+++ server/src/utils/types.ts
@@ -5,15 +5,6 @@ import type { FindGameError } from "../../../shared/types/api.ts";
 import { loadoutSchema } from "../../../shared/utils/loadout.ts";
 import type { MatchDataTable } from "../api/db/schema.ts";
 
-export interface GameSocketData {
-    gameId: string;
-    id: string;
-    closed: boolean;
-    rateLimit: Record<symbol, number>;
-    ip: string;
-    disconnectReason: string;
-}
-
 export const zUpdateRegionBody = z.object({
     regionId: z.string(),
     data: z.object({
@@ -40,16 +31,6 @@ export interface SaveGameBody {
 export interface ServerGameConfig {
     readonly mapName: keyof typeof MapDefs;
     readonly teamMode: TeamMode;
-}
-
-export interface GameData {
-    id: string;
-    teamMode: TeamMode;
-    mapName: string;
-    canJoin: boolean;
-    aliveCount: number;
-    startedTime: number;
-    stopped: boolean;
 }
 
 export const zFindGamePrivateBody = z.object({
@@ -79,84 +60,3 @@ export type FindGamePrivateRes =
         addrs: string[];
     }
     | { error: FindGameError };
-
-export enum ProcessMsgType {
-    Create,
-    Created,
-    KeepAlive,
-    UpdateData,
-    AddJoinToken,
-    SocketOpen,
-    ClientSocketMsg,
-    ServerSocketMsg,
-    SocketClose,
-}
-
-export interface CreateGameMsg {
-    type: ProcessMsgType.Create;
-    config: ServerGameConfig;
-    id: string;
-}
-
-export interface GameCreatedMsg {
-    type: ProcessMsgType.Created;
-}
-
-export interface KeepAliveMsg {
-    type: ProcessMsgType.KeepAlive;
-}
-
-export interface UpdateDataMsg extends GameData {
-    type: ProcessMsgType.UpdateData;
-}
-
-export interface AddJoinTokenMsg {
-    type: ProcessMsgType.AddJoinToken;
-    autoFill: boolean;
-    tokens: FindGamePrivateBody["playerData"];
-}
-
-export interface SocketOpenMsg {
-    type: ProcessMsgType.SocketOpen;
-    socketId: string;
-    ip: string;
-}
-
-export interface SocketClientMsg {
-    type: ProcessMsgType.ClientSocketMsg;
-    socketId: string;
-    data: ArrayBuffer | Uint8Array;
-}
-
-/**
- * msgs is an array to batch all msgs created in the same game net tick
- * into the same send call
- */
-export interface SocketServerMsg {
-    type: ProcessMsgType.ServerSocketMsg;
-    msgs: Array<{
-        socketId: string;
-        data: ArrayBuffer | Uint8Array;
-    }>;
-}
-
-/**
- * Sent by the server to the game when the socket is closed
- * Or by the game to the server when the game wants to close the socket
- */
-export interface SocketCloseMsg {
-    type: ProcessMsgType.SocketClose;
-    socketId: string;
-    reason?: string;
-}
-
-export type ProcessMsg =
-    | CreateGameMsg
-    | GameCreatedMsg
-    | KeepAliveMsg
-    | UpdateDataMsg
-    | AddJoinTokenMsg
-    | SocketOpenMsg
-    | SocketClientMsg
-    | SocketServerMsg
-    | SocketCloseMsg;
